% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Compute_total_loss}
\alias{Compute_total_loss}
\title{Compute the loss.}
\usage{
Compute_total_loss(Y, X)
}
\arguments{
\item{Y}{A \eqn{n} by \eqn{m} matrix.}

\item{X}{A \eqn{n} by \eqn{p} matrix.}
}
\value{
A list with two elements: \itemize{
  \item a vector of all losses,
  \item a vector of \eqn{d_j(PY)^2}.
}
}
\description{
Compute the loss \eqn{||Y-(PY)_r||^2} at different rank \eqn{r}.
}
