% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Est_rank}
\alias{Est_rank}
\title{Compute the rank.}
\usage{
Est_rank(losses, lambda, n, q, m, lowerRank = 0)
}
\arguments{
\item{losses}{The first element from \code{\link{Compute_total_loss}}.}

\item{lambda}{The penalty term.}

\item{n}{The number of rows of \code{data_Y}.}

\item{q}{The rank of \code{data_X}.}

\item{m}{The number of columns of \code{data_Y}.}

\item{lowerRank}{The smallest value of the target rank. The default is \eqn{0}.}
}
\value{
The estimated rank.
}
\description{
Compute the rank from the loss of \code{\link{Compute_total_loss}}.
}
