% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{MCSquarePEs}
\alias{MCSquarePEs}
\title{Approximate singualr values.}
\usage{
MCSquarePEs(q, m, Nsim = 200)
}
\arguments{
\item{q}{The rank of \code{data_X}.}

\item{m}{The number of columns of \code{data_Y}.}

\item{Nsim}{The number of repetitions in the MC with default equal to \eqn{200}.}
}
\value{
A vector of singular values from 1 to \code{q}.
}
\description{
Compute the singular values \eqn{d_j(PE)^2}, for j = 1, ..., \code{q},
by Monte Carlo simulations.
}
