% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_capitalize}
\alias{strs_capitalize}
\title{strs_capitalize}
\usage{
strs_capitalize(string, locale = "en")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{locale}{Locale to use for comparisons. See
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} for all possible options.
Defaults to "en" (English) to ensure that default behaviour is
consistent across platforms.}
}
\description{
Convert the first character of a string to uppercase and the
remaining characters to lowercase.
}
