% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{strs_contains}
\alias{strs_contains}
\title{strs_contains}
\usage{
strs_contains(string, substring)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\description{
Check whether a string contains the substring.
}
