% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{strs_find}
\alias{strs_find}
\title{strs_find}
\usage{
strs_find(string, substring)
}
\description{
Search each string for first occurence of \code{substring}. Returns
the substring index or NA if not found.
}
