% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isupper.R
\name{strs_isupper}
\alias{strs_isupper}
\title{strs_isupper}
\usage{
strs_isupper(string)
}
\description{
Check whether \code{string} only contains uppercase letters.
Technically this deviates from how it is defined in python which is that all
cased characters are uppercase with at least one cased character.
}
