% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len.R
\name{strs_len}
\alias{strs_len}
\title{Get the lengths of each string}
\usage{
strs_len(string)
}
\arguments{
\item{string}{A character vector of strings.}
}
\value{
A numeric vector where each element corresponds to the number of characters.
}
\description{
\code{strs_len} returns the number of characters in \code{string}. It is similar to
Python's \code{str.__len__()}.
}
\examples{
strs_len(c("hello", "world"))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#str.__len__}{Python str.\strong{len}() documentation}
}
