% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{strs_lower}
\alias{strs_lower}
\title{strs_lower}
\usage{
strs_lower(string, locale = "en")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{locale}{Locale to use for comparisons. See
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} for all possible options.
Defaults to "en" (English) to ensure that default behaviour is
consistent across platforms.}
}
\description{
strs_lower
}
