% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_rep}
\alias{strs_rep}
\title{strs_rep
Create a new string by repeating \code{string} \code{times}.}
\usage{
strs_rep(string, times)
}
\description{
strs_rep
Create a new string by repeating \code{string} \code{times}.
}
