% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_rfind}
\alias{strs_rfind}
\title{strs_rfind}
\usage{
strs_rfind(string, substring)
}
\description{
Search each string for the last occurence of \code{substring}.
Returns the substring index or NA if not found.
}
