% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapcase.R
\name{strs_swapcase}
\alias{strs_swapcase}
\title{Swap Uppercase and Lowercase Characters in a String}
\usage{
strs_swapcase(string)
}
\arguments{
\item{string}{A character vector where each element is a string.}
}
\value{
A character vector of the same length as \code{string}, with specified
uppercase characters converted to lowercase and visa-versa.
}
\description{
\code{strs_swapcase} returns a copy of the string with uppercase characters convert
to lowercase and visa-versa. It is similar to Python's \code{str.swapcase()}.
}
\examples{
strs_swapcase("Hello World")
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#str.swapcase}{Python str.swapcase() documentation}
}
