% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_textwrap}
\alias{strs_textwrap}
\title{strs_textwrap}
\usage{
strs_textwrap(
  string,
  width = 70,
  initial_indent = 0,
  subsequent_indent = 0,
  break_long_words = FALSE
)
}
\arguments{
\item{string}{character vector of strings to reformat.}

\item{width}{positive integer giving target line width in characters. A
width less than or equal to 1 will put each word on its own line.}

\item{initial_indent}{non-negative integer giving indentation of first line in
each paragraph}

\item{subsequent_indent}{non-negative integer giving indentation of following lines in
each paragraph}

\item{break_long_words}{If \code{FALSE}, the default, wrapping will only occur at
whitespace. If \code{TRUE}, can break on any non-word character (e.g. \code{/}, \code{-}).}
}
\value{
A character vector of re-wrapped strings.
}
\description{
Wrap strings into nicely formatted paragraphs.
This is a wrapper around \code{\link[stringi:stri_wrap]{stringi::stri_wrap()}} which implements
the Knuth-Plass paragraph wrapping algorithm.
}
