% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Update_lbd_DB}
\alias{Update_lbd_DB}
\title{Update lambda with deterministic bounds.}
\usage{
Update_lbd_DB(n, q, m, r, C = 0.01, simplified = F)
}
\arguments{
\item{n}{The number of rows of \code{data_Y}.}

\item{q}{The rank of \code{data_X}.}

\item{m}{The number of columns of \code{data_Y}.}

\item{r}{The selected rank.}

\item{C}{A numerical constant with default equal to \eqn{0.01}.}

\item{simplified}{A logical flag indicating whether or not using the simplified
deterministic bounds. The default is FALSE. It is only suitable to set
\code{simplified} to TRUE if n >> m or m << n.}
}
\value{
A numerical value of the new lambda.
}
\description{
Calculate the new lambda by using deterministic bounds.
}
