% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Update_lbd_MC}
\alias{Update_lbd_MC}
\title{Update lambda by MC.}
\usage{
Update_lbd_MC(n, q, m, r, SquarePEs, resid, C = 1.05)
}
\arguments{
\item{n}{The number of rows of \code{data_Y}.}

\item{q}{The rank of \code{data_X}.}

\item{m}{The number of columns of \code{data_Y}.}

\item{r}{The selected rank.}

\item{SquarePEs}{The sequence of singular values of PE computed from \code{\link{MCSquarePEs}}.}

\item{resid}{The residual value from MC simulations.}

\item{C}{A numerical constant with default equal to \eqn{1.05}.}
}
\value{
A numerical value of the new lambda.
}
\description{
Recalculate lambda based on the selected rank. Use Monte Carlo simulations
to approximate the value of \eqn{\frac{d_1^2(PE)+...+d_(2r)^2(PE)}{r}} at \code{r}.
}
