% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{The example matrix X}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 20 columns.
}
\usage{
X
}
\description{
\code{X} is a 50 by 20 numeric matrix.
}
\keyword{datasets}
