% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{The example matrix Y}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 20 columns.
}
\usage{
Y
}
\description{
\code{Y} is a 50 by 20 numeric matrix.
}
\keyword{datasets}
