% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{strs_center}
\alias{strs_center}
\title{strs_center}
\usage{
strs_center(string, width, fillchar = " ")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{width}{Minimum width of padded strings.}
}
\description{
Add padding to both sides of a string.
}
