% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_join}
\alias{strs_join}
\title{strs_join}
\usage{
strs_join(sep, ...)
}
\arguments{
\item{sep}{string inserted between each character}
}
\description{
If a single object is passed in after the separator, it's
assumed the user wants to flatten. If several objects are passed, it's
assumed that the user wants to concatenate.
}
