% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstrip.R
\name{strs_lstrip}
\alias{strs_lstrip}
\title{strs_lstrip}
\usage{
strs_lstrip(string, chars = NULL)
}
\description{
Remove \code{chars} from the left end of a string. If \code{chars} is
NULL, this function removes whitespace.
}
