% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{strs_replace}
\alias{strs_replace}
\title{strs_replace}
\usage{
strs_replace(string, substring, replacement)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{replacement}{The replacement value, usually a single string,
but it can be the a vector the same length as \code{string} or \code{pattern}.
References of the form \verb{\\1}, \verb{\\2}, etc will be replaced with
the contents of the respective matched group (created by \verb{()}).

Alternatively, supply a function, which will be called once for each
match (from right to left) and its return value will be used to replace
the match.}
}
\description{
Replace substrings in a string.
}
