% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_rstrip}
\alias{strs_rstrip}
\title{strs_rstrip}
\usage{
strs_rstrip(string, chars = NULL)
}
\description{
Remove \code{chars} from the right end of a string. If \code{chars} is
NULL, this function removes whitespace.
}
