% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{strs_split}
\alias{strs_split}
\title{strs_split}
\usage{
strs_split(string, sep = " ", maxsplit = -1L)
}
\description{
Split \code{string} on \code{char}. Passing a value for \code{n}
determines the number of splits to apply.
}
