% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_split}
\alias{strs_split}
\title{strs_split}
\usage{
strs_split(string, char = " ", n = Inf)
}
\description{
Split \code{string} on \code{char}. Passing a value for \code{n}
determines the number of splits to apply.
}
