% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs_strip}
\alias{strs_strip}
\title{strs_strip}
\usage{
strs_strip(string, chars = NULL)
}
\description{
Remove \code{chars} from both ends of a string. If \code{chars} is
NULL, this function removes whitespace.
}
